requirements.txt
scapy
streamlit
pandas
scikit-learn
numpy
plotly
Pillow
